# src/modules/auth/dto/admin-login.dto.ts
import { IsNotEmpty, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class AdminLoginDto {
  @ApiProperty({ description: 'Full name of the admin' })
  @IsString()
  @IsNotEmpty()
  fullName: string;

  @ApiProperty({ description: 'Password for the admin account' })
  @IsString()
  @IsNotEmpty()
  password: string;
}