# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login a user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent"
          }
        },
        "summary": "Request password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password successfully reset"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/social-login": {
      "post": {
        "operationId": "AuthController_socialLogin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User successfully logged in"
          },
          "400": {
            "description": "Invalid social login data"
          }
        },
        "summary": "Login via social media",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "operationId": "AuthController_adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin successfully logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Admin login with fullName and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the order"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/orders/admin/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all orders"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all orders (admin only)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/admin/orders/{id}/status": {
      "put": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order status updated successfully"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order status (admin only)",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully"
          },
          "409": {
            "description": "Post with this slug already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new blog post (admin only)",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Filter by tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of posts per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return published posts"
          }
        },
        "summary": "Get all published blog posts",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "operationId": "PostsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the post"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "summary": "Get a blog post by slug",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "put": {
        "operationId": "PostsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully"
          },
          "404": {
            "description": "Post not found"
          },
          "409": {
            "description": "Post with this slug already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a blog post (admin only)",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a blog post (admin only)",
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Đông Y Pharmacy API",
    "description": "The Đông Y Pharmacy REST API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number of the user"
          },
          "password": {
            "type": "string",
            "description": "Password for the user account"
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user"
          },
          "password": {
            "type": "string",
            "description": "Password for the user account"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Reset password token"
          },
          "password": {
            "type": "string",
            "description": "New password for the user account"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Full name of the admin"
          },
          "password": {
            "type": "string",
            "description": "Password for the admin account"
          }
        },
        "required": [
          "fullName",
          "password"
        ]
      },
...
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New order status",
            "enum": [
              "pending",
              "processing",
              "shipped",
              "delivered",
              "cancelled"
            ]
          },
          "trackingNumber": {
            "type": "string",
            "description": "Tracking number for shipped orders"
          },
          "cancelReason": {
            "type": "string",
            "description": "Reason for cancellation if status is cancelled"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the blog post"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the post"
          },
          "content": {
            "type": "string",
            "description": "Full content of the post in HTML/Markdown"
          },
          "summary": {
            "type": "string",
            "description": "Short summary of the post"
          },
          "imageUrl": {
            "type": "string",
            "description": "URL to the featured image"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether the post is published",
            "default": true
          },
          "tags": {
            "description": "Tags associated with the post",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "authorName": {
            "type": "string",
            "description": "Name of the post author"
          }
        },
        "required": [
          "title",
          "slug",
          "content"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the blog post"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug for the post"
          },
          "content": {
            "type": "string",
            "description": "Full content of the post in HTML/Markdown"
          },
          "summary": {
            "type": "string",
            "description": "Short summary of the post"
          },
          "imageUrl": {
            "type": "string",
            "description": "URL to the featured image"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether the post is published"
          },
          "tags": {
            "description": "Tags associated with the post",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "authorName": {
            "type": "string",
            "description": "Name of the post author"
          }
        }
      },
...
    }
  }
}